refactor(rpc): move context into own file
